requirements.txt
asgiref==3.6.0
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.1.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.2
daphne==4.0.0
defusedxml==0.7.1
dj-database-url==2.0.0
Django==4.2
django-cors-headers==4.0.0
django-filter==23.2
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
google-auth==2.19.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.4
importlib-metadata==6.6.0
incremental==22.10.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.4.3
MarkupSafe==2.1.2
oauthlib==3.2.2
packaging==23.1
Pillow==9.5.0
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==23.1.1
python-decouple==3.8
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
service-identity==21.1.0
shortuuid==1.0.11
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
Twisted==22.10.0
# twisted-iocpsupport==1.0.3
txaio==23.1.1
typing_extensions==4.5.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
zipp==3.15.0
zope.interface==6.0
